moved the CSS files into a CSS folder
