+++ mai_weatherApp/client/src/Home.jsx
@@ -2,7 +2,7 @@ import CreateUser from './CreateUser';
 import WeatherAPI from './WeatherAPI';
 import { UserProvider } from './UserContext';
 import User from './User';
-import './Home.css';
+import './CSS/Home.css';
 
 function Home() {
   return (
