+++ mai_weatherApp/client/src/main.jsx
@@ -1,6 +1,6 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import './index.css'
+import './CSS/index.css'; 
 import Home from './Home.jsx'
 
 createRoot(document.getElementById('root')).render(
